+++ src/backend/joanie/core/models/products.py
@@ -695,7 +695,7 @@ class Order(BaseModel):
                     )
                 )
             if (
-                not self.created_on or self.state not in [enums.ORDER_STATE_CANCELED]
+                not self.created_on and self.state not in [enums.ORDER_STATE_CANCELED]
             ) and (
                 self.enrollment.course_run.course.state["priority"]
                 >= CourseState.ARCHIVED_CLOSED